Write to the ReusableBuffer on-demand rather than in advance
If there are a lot of sounds to decode on program load, that can take a
while. Instead, only populate the buffer Vec once the internal iterator
has been read to defer that work until it’s needed.

This does unfortunately mean we need a RwLock whereas previously we
didn’t.