Use double precision for IntoSampleRate::position
If a sound plays at 96KHz, we quickly run out of precision when storing
position in a 32-bit float. This was causing a bug where the pitch would
change occasionally due to precision errors.

A f64 should be sufficient based on these calculations:
  - A change of 1Hz is 96000/95999 = 1.0000104168
  - We need ~17 bits of precision to represent this fractional part
  - That leaves 47 bits for the whole number part
  - 2^(47 - 1) / 96000 = 733,007,752
  - 733,007,752 seconds is ~23 years of playback

I think that means we could play an audio file for 23 years before
running into precision errors with the f64 - although I’m not 100% sure.