Add a top-level AudioMixer::pause method
In my game, I’m wrapping all sources in PauseWhenMuted solely for the
purpose of pausing audio when the in-game menu is visible. Because this
iterator runs all the time, it adds overhead to every audio source and
99% of the time it won’t be useful since the game won’t be paused.

Therefore, introduce a top-level AudioMixer::pause method that handles
this use case more efficiently. It still has some overhead but it should
be negligible compared to the previous approach.

We still need to play sounds when the in-game menu is open (the click)
so make it so new sources can be added when AudioMixer is paused and
play them as normal.